video_chat2/requirements.txt
apex==0.9.10dev
av==10.0.0
decord==0.6.0
einops==0.6.1
flash_attn==1.0.4
fvcore==0.1.5.post20221221
gradio==3.29.0
imageio==2.27.0
iopath==0.1.10
mmcv==2.0.0
numpy==1.23.5
omegaconf==2.3.0
opencv_python==4.7.0.72
pandas==1.5.3
Pillow==9.5.0
psutil==5.9.4
PyYAML==6.0
PyYAML==6.0
scipy==1.10.1
termcolor==2.3.0
timm==0.6.12
torch==1.13.1+cu117
torchvision==0.14.1+cu117
tqdm==4.64.1
transformers==4.28.1
sentencepiece==0.1.99
wandb==0.14.0
peft==0.3.0
